requirements.conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
aiohttp=3.8.1=py38h7f8727e_1
aiosignal=1.2.0=pyhd3eb1b0_0
anyio=3.5.0=py38h06a4308_0
anykeystore=0.2=pypi_0
apex=0.1=pypi_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py38h7f8727e_0
async-timeout=4.0.1=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
autopep8=1.6.0=pypi_0
axial-attention=0.6.1=pypi_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
black=22.6.0=py38h06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
boto3=1.20.24=pyhd3eb1b0_0
botocore=1.23.24=pyhd3eb1b0_0
brotli=1.0.9=h7f98852_5
brotli-bin=1.0.9=h7f98852_5
brotlipy=0.7.0=py38h27cfd23_1003
c-ares=1.18.1=h7f8727e_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
catalogue=2.0.6=py38h06a4308_0
certifi=2023.5.7=py38h06a4308_0
cffi=1.15.0=py38hd667e15_1
charset-normalizer=2.0.9=pypi_0
click=8.0.3=pypi_0
colorama=0.4.4=pyhd3eb1b0_0
cryptacular=1.6.2=pypi_0
cryptography=36.0.0=py38h9ce1e76_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.11.0=pyhd8ed1ab_0
cymem=2.0.5=py38h2531618_0
cython-blis=0.7.4=py38h27cfd23_1
dataclasses=0.8=pyh6d0b6a4_7
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py38h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dgl=0.6.1=pypi_0
docker-pycreds=0.4.0=pypi_0
en-core-web-sm=3.2.0=pypi_0
entrypoints=0.3=py38_0
expat=2.4.4=h295c915_0
filelock=3.4.0=pypi_0
flake8=6.0.0=py38h06a4308_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=py38h7f8727e_0
giflib=5.2.1=h7b6447c_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
glib=2.69.1=h4ff587b_1
google-api-core=1.25.1=pyhd3eb1b0_0
google-auth=1.33.0=pyhd3eb1b0_0
google-cloud-core=1.7.1=pyhd3eb1b0_0
google-cloud-storage=1.41.0=pyhd3eb1b0_0
google-crc32c=1.1.2=py38h27cfd23_0
google-resumable-media=1.3.1=pyhd3eb1b0_1
googleapis-common-protos=1.53.0=py38h06a4308_0
greenlet=1.1.2=pypi_0
grpcio=1.42.0=py38hce63b2e_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
huggingface-hub=0.2.1=pypi_0
hupper=1.10.3=pypi_0
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.8.2=py38h06a4308_0
importlib_metadata=4.8.2=hd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.4.1=py38h06a4308_1
ipython=7.31.1=py38h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py38h06a4308_1
jinja2=3.0.2=pyhd3eb1b0_0
jmespath=0.10.0=pyhd3eb1b0_0
joblib=1.1.0=pypi_0
jpeg=9b=h024ee3a_2
json5=0.9.6=pyhd3eb1b0_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter=1.0.0=py38_7
jupyter-http-over-ws=0.0.8=pypi_0
jupyter_client=7.1.2=pyhd3eb1b0_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.9.1=py38h06a4308_0
jupyter_server=1.13.5=pyhd3eb1b0_0
jupyterlab=3.2.9=pyhd3eb1b0_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.10.3=pyhd3eb1b0_1
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.2=py38h295c915_0
langcodes=3.3.0=pyhd3eb1b0_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libbrotlicommon=1.0.9=h7f98852_5
libbrotlidec=1.0.9=h7f98852_5
libbrotlienc=1.0.9=h7f98852_5
libcrc32c=1.1.1=he6710b0_2
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libpng=1.6.37=hbc83047_0
libprotobuf=3.19.1=h4ff587b_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.0.1=py38h27cfd23_0
matplotlib=3.5.1=py38h578d9bd_0
matplotlib-base=3.5.1=py38ha18d171_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mccabe=0.7.0=pyhd3eb1b0_0
mistune=0.8.4=py38h7b6447c_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
multidict=5.2.0=py38h7f8727e_2
munkres=1.0.7=py_1
murmurhash=1.0.5=py38h2531618_0
mypy_extensions=0.4.3=py38h06a4308_1
nbclassic=0.3.5=pyhd3eb1b0_0
nbclient=0.5.11=pyhd3eb1b0_0
nbconvert=6.3.0=py38h06a4308_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.6.3=pypi_0
ninja=1.10.2=py38hd09550d_3
nltk=3.7=pypi_0
notebook=6.4.8=py38h06a4308_0
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
nvidia-htop=1.0.5=pypi_0
oauthlib=3.2.0=pypi_0
olefile=0.46=pyhd3eb1b0_0
openssl=1.1.1t=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.2.3=py38h51da96c_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pastedeploy=2.1.1=pypi_0
pathspec=0.10.3=py38h06a4308_0
pathtools=0.1.2=pypi_0
pathy=0.6.0=pyhd3eb1b0_0
pbkdf2=1.3=pypi_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py38h5aabda8_0
pip=21.2.4=py38h06a4308_0
plaster=1.0=pypi_0
plaster-pastedeploy=0.7=pypi_0
platformdirs=2.5.2=py38h06a4308_0
preshed=3.0.5=py38h2531618_4
prometheus_client=0.13.1=pyhd3eb1b0_0
promise=2.3=pypi_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
protobuf=3.19.1=py38h295c915_0
psutil=5.9.0=pypi_0
ptflops=0.6.9=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycodestyle=2.8.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.8.2=py38h7f8727e_0
pyflakes=3.0.1=py38h06a4308_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.6=pypi_0
pyqt=5.9.2=py38h05f1152_4
pyramid=2.0=pypi_0
pyramid-mailer=0.15.1=pypi_0
pyrsistent=0.18.0=py38heee7806_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python3-openid=3.2.0=pypi_0
python_abi=3.8=2_cp38
pytorch=1.7.1=py3.8_cuda11.0.221_cudnn8.0.5_0
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=py38h295c915_2
qt=5.9.7=h5867ecd_1
qtconsole=5.2.2=pyhd3eb1b0_0
qtpy=1.11.2=pyhd3eb1b0_0
readline=8.1=h27cfd23_0
regex=2021.11.10=pypi_0
repoze-sendmail=4.4.1=pypi_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rouge=1.0.1=pypi_0
rsa=4.7.2=pyhd3eb1b0_1
s3transfer=0.5.0=pyhd3eb1b0_0
sacremoses=0.0.46=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
sentence-transformers=2.2.0=pypi_0
sentencepiece=0.1.96=pypi_0
sentry-sdk=1.5.10=pypi_0
setproctitle=1.2.3=pypi_0
setuptools=58.0.4=py38h06a4308_0
shellingham=1.3.1=pyhd3eb1b0_0
shortuuid=1.0.8=pypi_0
sip=4.19.13=py38h295c915_0
six=1.16.0=pyhd3eb1b0_0
sklearn=0.0=pypi_0
smart_open=5.1.0=pyhd3eb1b0_0
smmap=5.0.0=pypi_0
sniffio=1.2.0=py38h06a4308_1
spacy=3.2.1=py38hae6d005_0
spacy-legacy=3.0.8=pyhd3eb1b0_0
spacy-loggers=1.0.1=pyhd3eb1b0_0
sqlalchemy=1.4.32=pypi_0
sqlite=3.36.0=hc218d9a_0
srsly=2.4.1=py38h2531618_0
tensorboardx=2.5=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.13.1=py38h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
thinc=8.0.13=py38hae6d005_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.11=h1ccaba5_0
tokenizers=0.10.3=pypi_0
toml=0.10.2=pypi_0
tomli=2.0.1=py38h06a4308_0
torchaudio=0.7.2=py38
torchvision=0.8.2=py38_cu110
tornado=6.1=py38h27cfd23_0
tqdm=4.62.3=pyhd3eb1b0_1
traitlets=5.1.1=pyhd3eb1b0_0
transaction=3.0.1=pypi_0
transformers=4.14.1=pypi_0
translationstring=1.4=pypi_0
typer=0.4.0=pyhd3eb1b0_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
ujson=5.1.0=py38h295c915_0
urllib3=1.26.7=pypi_0
velruse=1.1.1=pypi_0
venusian=3.0.0=pypi_0
wandb=0.12.16=pypi_0
wasabi=0.8.2=pyhd3eb1b0_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
webob=1.8.7=pypi_0
websocket-client=0.58.0=py38h06a4308_4
wheel=0.37.0=pyhd3eb1b0_1
widgetsnbextension=3.5.2=py38h06a4308_0
wtforms=3.0.1=pypi_0
wtforms-recaptcha=0.3.2=pypi_0
xz=5.2.5=h7b6447c_0
yarl=1.5.1=py38h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.11=h7f8727e_4
zope-deprecation=4.4.0=pypi_0
zope-interface=5.4.0=pypi_0
zope-sqlalchemy=1.6=pypi_0
zstd=1.4.9=haebb681_0
